Rename ErrorContext to MessageContext to contain more information
