+++ ui/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import { DataProvider } from "./context/DataContext";
-import { ErrorProvider } from "./context/ErrorContext";
+import { MessageProvider } from "./context/MessageContext";
 // import "./globals.css";
 
 const inter = Inter({ subsets: ["latin"] });
@@ -19,11 +19,11 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body className={inter.className}>
-        <ErrorProvider>
+        <MessageProvider>
           <DataProvider>
             {children}
           </DataProvider>
-        </ErrorProvider>
+        </MessageProvider>
       </body>
     </html>
   );
