+++ ui/app/results/ResultGrid.tsx
@@ -3,7 +3,7 @@ import { DataGrid, GridColDef, GridRowsProp, GridToolbar } from '@mui/x-data-gri
 import { Paper, Switch, FormControlLabel, Box } from '@mui/material';
 import InstructionBox from './InstructionBox';
 import { useData } from '../context/DataContext';
-import { useError } from '../context/ErrorContext';
+import { useError } from '../context/MessageContext';
 import { useLoading } from '../context/LoadingContext';
 import { useRouter } from 'next/navigation';
 import { searchPapers } from '../actions';
